requirements.txt
sentence-transformers>=2.7.0
transformers>=4.40.0
torch>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
tqdm>=4.66.0
psycopg2-binary>=2.9.9
pgvector>=0.3.0
